+++ CHANGELOG.md
@@ -7,10 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 * Created a vignette covering the testing of histograms (@ryla5068, #149)
-
-## [0.1.0]
 * Added test for bin heights of histograms (@ryla5068, #124)
 * Added support for overlapping histograms in histogram tests (@ryla5068, #123)
+
+## [0.1.0]
 * Created a vignette covering base plot tester functionality (@ryla5068, #122)
 * fix pip version to ensure pyproj installs in black tox envt (@lwasser, #144)
 * Changed `get_caption()` to return a string (@ryla5068, #125)